Refactor all shaders to jsx functions
Get rid of the use of glsl-template-loader, as glsl-man is buggy.  Also
by removing dependance on glsl-man and the template loader, template
literals can be taken advantage of, for example loading editor programs
into an existing shader is not as strange.  Finally RawShaderMaterial is
now used for the compute shader.